+++ controllers/zuul.go
@@ -190,12 +190,6 @@ func (r *SFController) mkZuulContainer(service string, corporateCMExists bool) a
 			Name:      "zuul-ca",
 			MountPath: "/etc/pki/ca-trust/extracted",
 		},
-		{
-			Name:      "extra-python-module",
-			MountPath: "/usr/local/lib/python3.11/site-packages/sfExtras.py",
-			SubPath:   "sfExtras.py",
-			ReadOnly:  true,
-		},
 	}
 	envs := []apiv1.EnvVar{
 		base.MkEnvVar("REQUESTS_CA_BUNDLE", "/etc/ssl/certs/ca-bundle.crt"),
@@ -300,11 +294,6 @@ func mkZuulVolumes(service string, r *SFController, corporateCMExists bool) []ap
 
 	// Install the logging settings config map resource
 	r.EnsureConfigMap("zuul-logging", r.computeLoggingConfig())
-
-	// Install extra python module
-	r.EnsureConfigMap("extra-python-module", map[string]string{
-		"sfExtras.py": logging.SFExtrasPythonModule,
-	})
 
 	volumes := []apiv1.Volume{
 		base.MkVolumeSecret("ca-cert"),
@@ -470,7 +459,6 @@ func (r *SFController) EnsureZuulScheduler(cfg *ini.File) bool {
 		"zuul-extra":                 utils.Checksum([]byte(sshConfig)),
 		"zuul-connections":           utils.IniSectionsChecksum(cfg, utils.IniGetSectionNamesByPrefix(cfg, "connection")),
 		"corporate-ca-certs-version": getCMVersion(corporateCM, corporateCMExists),
-		"extras-python-module":       utils.Checksum([]byte(logging.SFExtrasPythonModule)),
 	}
 
 	if r.isConfigRepoSet() {
@@ -598,7 +586,6 @@ func (r *SFController) EnsureZuulExecutor(cfg *ini.File) bool {
 		"zuul-logging":               utils.Checksum([]byte(r.getZuulLoggingString("zuul-executor"))),
 		"zuul-connections":           utils.IniSectionsChecksum(cfg, utils.IniGetSectionNamesByPrefix(cfg, "connection")),
 		"corporate-ca-certs-version": getCMVersion(corporateCM, corporateCMExists),
-		"extras-python-module":       utils.Checksum([]byte(logging.SFExtrasPythonModule)),
 	}
 	// TODO Add the zk-port-forward-kube-config secret resource version in the annotation if enabled
 
@@ -676,7 +663,6 @@ func (r *SFController) EnsureZuulMerger(cfg *ini.File) bool {
 		"zuul-connections":           utils.IniSectionsChecksum(cfg, utils.IniGetSectionNamesByPrefix(cfg, "connection")),
 		"zuul-logging":               utils.Checksum([]byte(r.getZuulLoggingString("zuul-merger"))),
 		"corporate-ca-certs-version": getCMVersion(corporateCM, corporateCMExists),
-		"extras-python-module":       utils.Checksum([]byte(logging.SFExtrasPythonModule)),
 	}
 
 	zm := r.mkHeadlessSatefulSet(service, "", r.getStorageConfOrDefault(r.cr.Spec.Zuul.Merger.Storage), apiv1.ReadWriteOnce, r.cr.Spec.ExtraLabels)
@@ -742,7 +728,6 @@ func (r *SFController) EnsureZuulWeb(cfg *ini.File) bool {
 		"zuul-logging":               utils.Checksum([]byte(r.getZuulLoggingString("zuul-web"))),
 		"zuul-connections":           utils.IniSectionsChecksum(cfg, utils.IniGetSectionNamesByPrefix(cfg, "connection")),
 		"corporate-ca-certs-version": getCMVersion(corporateCM, corporateCMExists),
-		"extras-python-module":       utils.Checksum([]byte(logging.SFExtrasPythonModule)),
 	}
 
 	zw := base.MkDeployment("zuul-web", r.ns, "", r.cr.Spec.ExtraLabels)